move types to new typing module
