+++ temporian/__init__.py
@@ -57,7 +57,6 @@ from temporian.core.data import duration
 
 # EventSets
 from temporian.implementation.numpy.data.event_set import EventSet
-from temporian.implementation.numpy.data.event_set import EventSetOrNode
 from temporian.implementation.numpy.data.io import event_set
 
 # Serialization
@@ -81,6 +80,9 @@ from temporian.implementation.numpy.data.plotter import plot
 
 # Compilation
 from temporian.core.compilation import compile
+
+# Types
+from temporian.core.typing import EventSetOrNode
 
 
 # --- OPERATORS ---
