+++ temporian/core/operators/binary/BUILD
@@ -38,6 +38,7 @@ py_library(
         ":base",
         "//temporian/core:compilation",
         "//temporian/core:operator_lib",
+        "//temporian/core:typing",
         "//temporian/core/data:dtype",
         "//temporian/core/data:node",
     ],
