+++ temporian/core/operators/binary/arithmetic.py
@@ -19,7 +19,7 @@ from temporian.core.compilation import compile
 from temporian.core.data.node import EventSetNode
 from temporian.core.data.dtype import DType
 from temporian.core.operators.binary.base import BaseBinaryOperator
-from temporian.implementation.numpy.data.event_set import EventSetOrNode
+from temporian.core.typing import EventSetOrNode
 
 
 class BaseArithmeticOperator(BaseBinaryOperator):
